+++ test/cljstyle/config_test.clj
@@ -141,7 +141,7 @@
           (is (not (config/ignored? config #{} foo-clj)))
           (is (config/ignored? config #{} (io/file test-dir "foo")))
           (is (config/ignored? config #{} (io/file test-dir "bar")))
-          (is (config/ignored? {} #{"**/bar"} (io/file test-dir "bar")))))
+          (is (config/ignored? {} #{"bar"} (io/file test-dir "bar")))))
       (finally
         (when (.exists foo-clj)
           (.delete foo-clj))))))